moved definitions of HITRAN groups and list of supported molecules in radis/db
